New Setup: Amount for Asset Type Insurances
For insurances (Asset Typ enum: life_insurance_3a, life_insurance_3b) two different types of amount are interesting for the FI:

surrender value / Rückkaufswert
insurance sum / Versicherungssumme
Both of them will be entered in the core banking system. So we would like to transfer both amounts.
Is it possible to add a new field in the object "Asset":
name: insurance sum
description: The amount of the sum insured
Type: see Amount

So, we can send the surrender value in the current field "amount" and send the insurance sum in the new field.