Required Fields in Liabilities and Income
In "Asset" the fields "assetType" and "amount" are required. in "liabilites" and "income", these two fields are not required. How come? We should mark these two fields in "liabilities" and "income" as required, too. -> added "required" for both

There is one more difference: asset has an additional field "assetProvider". Do we need/want this field for "liability" and "income" as well? -> added this additional information fields.

Anyway, why is there a different specification format for "liabilities" and "income" than for "asset"? "asset" is separately specified under "mortgage date types" whereas the other two are specified in the normal structure. Do we want to change it? -> didn't change. It has a reason for different approach. Asset is used in different objects and should therefore be specified in a separate structure. that's good.