Amortizations as an own object
We need the data fields with connection to amortization not only in the object "financingTranche" but also on a more aggregate level in the object "financing". Reason: Not always the amortization is already fixed on a tranche level. Sometimes the amortization is aggred on a total mortgage level.

After this change, we have amortization at three different places: financing, financingtranche and mortgageproductcondition. According to the new style guideline, using references is recommended. So, we want to clean up the amortization documentation.